app/requirements.txt
git+https://github.com/Lawly-code/database.git@0.2.11#egg=lawly-db
fastapi==0.115.12
pre-commit==4.2.0
PyJWT==2.10.1
pydantic_settings==2.9.1
uvicorn==0.34.2
email-validator == 2.2.0
